chore: Move dev dependencies into dev dependencies
I noticed these deps were being installed when I was working with the
SDK recently.

I think [`save-dev`](https://www.npmjs.com/package/save-dev) was added
due to an errant npm i save-dev since it is a holding package so I
removed it (sorry if I added it haha!).

`@types/dotenv` is no longer needed since the `dotenv` package pulls
this in independently. It was pulling on `dotenv` as a peerDep (I think)
so this just moves to depending on the package itself.

The typedoc generator is also used only for development / ci (as far as
I know) so moving.